Develop

See merge request s05-webmobile1-sub3/S05P13A607!113

+++ backend/src/main/java/com/ssafy/booktory/util/S3Uploader.java
@@ -6,12 +6,16 @@ import com.amazonaws.services.s3.model.PutObjectRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.io.Resource;
+import org.springframework.core.io.ResourceLoader;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStream;
+import java.nio.file.Path;
 import java.sql.Date;
 import java.time.LocalDate;
 import java.util.Arrays;
@@ -23,7 +27,7 @@ import java.util.UUID;
 @Component
 public class S3Uploader implements Uploader{
 
-    private final static String TEMP_FILE_PATH = "src/main/resources/static/";
+    private final static String TEMP_FILE_PATH = "classpath:";
     private final AmazonS3Client amazonS3Client;
     private final List<String> imageExt = Arrays.asList(".PNG", ".png", ".JPEG", ".jpeg", ".JPG", ".jpg", ".TIFF", ".tiff", ".BMP", ".bmp" , ".GIF", ".gif");
 
@@ -42,7 +46,9 @@ public class S3Uploader implements Uploader{
 
     @Override
     public void deleteS3Instance(String url) {
-        System.out.println(s3Url);
+        log.info("file url : " + url);
+        if(url.indexOf(s3Url) == -1)
+            return ;
         String fileName = url.replaceFirst(s3Url, "");
         amazonS3Client.deleteObject(bucket, fileName);
         log.info(fileName + " 삭제완료");
@@ -67,6 +73,7 @@ public class S3Uploader implements Uploader{
 
     private File convert(MultipartFile file) throws IOException{
         File convertFile = new File(TEMP_FILE_PATH + file.getOriginalFilename());
+
         if(convertFile.createNewFile()){
             try(FileOutputStream fos = new FileOutputStream(convertFile)){
                 fos.write(file.getBytes());
